requirements.txt
aiogram>=3.10,<4
aiosqlite>=0.20
python-dotenv>=1.0
nodriver>=0.38
